+++ mtg_proxy_printer/carddb_migrations.py
@@ -104,9 +104,7 @@ class Migrate21to22(MigrationScript):
                 "q": f"date>1970-01-01 date<={timestamp.date()}"
             })
             try:
-                card_count = next(aw.read_json_card_data_from(
-                    f'https://api.scryfall.com/cards/search?{url_parameters}', 'total_cards'
-                ))
+                card_count = next(aw.get_available_card_count())
             except (urllib.error.URLError, socket.error):
                 card_count = 0
             data.append((id_, timestamp.isoformat(), card_count))