+++ doc/changelog.md
@@ -7,10 +7,9 @@
 - MTG sets are now shown with their set symbols wherever a set can be selected.
 - Configurable printing preferences:
   - Add ability to specify positive or negative weights for card properties, like border color, etc.,
-    and individual Magic sets.
+    and for individual Magic sets.
   - The app uses these to calculate the best matching printing whenever it has to choose among multiple choices.
-  - Negative values will rank matching printings lower than plain, regular printings, so they will only be chosen,
-    if there is no alternative available.
+  - Negative values will rank matching printings lower, so they will only be chosen, if there is no alternative available.
 - New printing filters:
   - For Universes Beyond cards (A negative preference weight for UB cards can be used to avoid UB-reprints of in-universe cards.) 
   - For cards with low resolution images