+++ doc/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Next version (in development)
+# Version 0.36.0 (2026-07-31)  <a name="v0_36_0"></a>
 
 ## New features